requirements.txt
celery==5.2.7
coverage==7.2.1
Django==4.1.7
django-celery-beat==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
ipython==8.10.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1